+++ geoimagenet_ml/api/routes/models/views.py
@@ -1,6 +1,6 @@
 from geoimagenet_ml.api.routes.models.utils import create_model, get_model
 from geoimagenet_ml.api import exceptions as ex, schemas as s
-from geoimagenet_ml.store.constants import OPERATION
+from geoimagenet_ml.constants import OPERATION
 from geoimagenet_ml.store.datatypes import Model
 from geoimagenet_ml.store.factories import database_factory
 from pyramid.response import FileResponse
@@ -18,7 +18,7 @@ def get_models_view(request):
                                    fallback=lambda: db.rollback(), httpError=HTTPInternalServerError,
                                    request=request, msgOnFail=s.InternalServerErrorResponseSchema.description)
     db.actions_store.save_action(Model, OPERATION.LIST, request=request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'models': models_json},
+    return ex.valid_http(httpSuccess=HTTPOk, content={u"models": models_json},
                          detail=s.Models_GET_OkResponseSchema.description, request=request)
 
 
@@ -27,7 +27,7 @@ def post_models_view(request):
     """Register a new model."""
     model = create_model(request)
     database_factory(request).actions_store.save_action(model, OPERATION.UPLOAD, request=request)
-    return ex.valid_http(httpSuccess=HTTPCreated, content={u'model': model.json()},
+    return ex.valid_http(httpSuccess=HTTPCreated, content={u"model": model.json()},
                          detail=s.Models_POST_CreatedResponseSchema.description, request=request)
 
 
@@ -36,9 +36,16 @@ def post_models_view(request):
 def get_model_view(request):
     """Get registered model information."""
     model = get_model(request)
-    database_factory(request).actions_store.save_action(model, OPERATION.INFO, request=request)
-    return ex.valid_http(httpSuccess=HTTPOk, content={u'model': model.json()},
-                         detail=s.Models_GET_OkResponseSchema.description, request=request)
+    db = database_factory(request)
+    db.actions_store.save_action(model, OPERATION.INFO, request=request)
+    _, download_count = db.actions_store.find_actions(item_type=Model, item=model.uuid, operation=OPERATION.DOWNLOAD)
+    model_content = {
+        "model": model.json(),
+        "owner": model.user,
+        "downloads": download_count,
+    }
+    return ex.valid_http(httpSuccess=HTTPOk, content=model_content,
+                         detail=s.Model_GET_OkResponseSchema.description, request=request)
 
 
 @s.ModelDownloadAPI.get(tags=[s.ModelsTag],